Allow the library to work as a merged component
Fixes as issue created if SharpKml is merged into an assembly that
depends on it. In this case 'assembly' refers to the assembly using
SharpKml. The element types will not resolve correctly unless they are
included in the publicly exported types of dependent assembly. This
small change means that all types will be returned from the assembly,
but will still be filtered for sub-classes of Element, hence
RegisterAssembly performs as designed, but in the merged assembly
environment.